requirements.txt
-i https://pypi.org/simple
aiohttp==3.7.4.post0
asgiref==3.3.1; python_version >= '3.5'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-filter==2.4.0
django==3.1.7
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
drf-spectacular==0.14.0
freezegun==1.1.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1; python_version >= '3.5'
jsonschema==3.2.0
multidict==5.1.0; python_version >= '3.6'
psycopg2==2.8.6
pyjwt==2.0.1; python_version >= '3.6'
pyrsistent==0.17.3; python_version >= '3.5'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.25.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
typing-extensions==3.7.4.3
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
yarl==1.6.3; python_version >= '3.6'
